[Refactor] Do not rely on implicit conversions in {fmt} calls  (#72022)

+++ be/src/util/compression/stream_decompressor.cpp
@@ -1152,7 +1152,7 @@ StatusOr<std::unique_ptr<StreamDecompressor>> StreamDecompressor::create_decompr
         decompressor = std::make_unique<LzoStreamDecompressor>();
         break;
     default:
-        return Status::InternalError(fmt::format("Unknown compress type: {}", type));
+        return Status::InternalError(fmt::format("Unknown compress type: {}", static_cast<int>(type)));
     }
 
     RETURN_IF_ERROR(decompressor->init());
